refactor(card): 카드 키, 몸무게 타입 수정

+++ src/main/java/com/dekk/card/application/command/CardCreateCommand.java
@@ -11,7 +11,7 @@ public record CardCreateCommand(
     String originId,
     boolean isActive,
     Platform platform,
-    Double height,
-    Double weight
+    Integer height,
+    Integer weight
 ) {
 }
